Tray icon and opening by shortcut functionality (#1)
* feat(totes): implement try icon functionallity.

ALl functionallity is added based on this guide:
* https://tauri.app/v1/guides/features/system-tray

Currently it works well on Windows. I haven't tested it on Linux.

* feat(totes): add `rustdmt.toml` files;

* feat(totes): implement opening and hiding app on shortcut.

The implementation is based on this answer:
https://github.com/tauri-apps/tauri/discussions/5281#discussioncomment-3733845.
The shortcut itself if hardcoded and can't be configured. We plan to
change it in the future.
It works perfectly on Windows. I haven't tested it on Linux.