feat(front): implement list autocontinue when typing a new line; (#123)
This PR implements automatic list incrementing when typing a new line. It works for both types of lists: unordered and ordered.

There is a small limitation: **list items must be one-line text**. If the list item consists of more than one line of text, then it will not work. It is done on purpose. It is hard to parse the previous list item. It can be just text, or any other Markdown component (table, code block, etc.). We do not want to do MD parsing on the frontend.

The primary purpose of this feature is to assist users in the most common scenarios when writing lists.