Fix synchronization algorithm in many places (#156)
This PR fixes a bunch of bugs in the synchronization algorithm. There is a short description:

## Fixed remote operations applying

Sometimes we used to get the `sqlx` error about a non-existent row. It was happening because the app was searching only in notes/spaces with `is_deleted = FALSE`. It was wrong because it is possible to update the note on one device after it was deleted on another device.

If the note is updated after its deletion, then we should restore it and apply the changes.

## Fixed the `CreateFile` operation applying

Previously, we assumed that when we accept the `CreateFile` operation, it also means that the file is uploaded. But it is not necessarily true. The operation can be synced with the server before the actual file upload happens. For example, when file uploading fails but the operations uploading succeed, or when two concurrent sync processes are running.

## Improved file synchronization algorithm

Now, when the file is not uploaded and does not exist locally, we try to download it from the sync server anyway (see bug description above). Sometimes it can help. Otherwise, we just continue.

## Fixed the common operations elimination algorithm

At step 4 of the synchronization algorithm (https://github.com/TheBestTvarynka/Dataans/blob/e0b80622160b718b122dcd7b5cea0491e27b2a5c/dataans/src-tauri/src/dataans/sync/mod.rs#L46-L48), the app has lists of remote and local operations and needs to eliminate common operations from these lists.
Previously, the app compared IDs from two lists one by one, and stopped on two operations with different IDs. The assumption was that starting from this point, the app does not need to check other operations in lists, because it is guaranteed that their IDs will be different.

But, unfortunately, it is not true, so I reimplemented this part (https://github.com/TheBestTvarynka/Dataans/pull/156/commits/07d6a6244695ee37657fa41ef64516ca9d492010). I decided to keep it simple and used `HashSet` from `sts` and its `.intersection` method. Yes, it can be better and can be optimized (for example, using the [Longest Common Subsequence (LCS)](https://en.wikipedia.org/wiki/Longest_common_subsequence) algorithm), but I do not need it for now.

## Allow operation reuploading and reapplying existing operations

Now the user is allowed to upload the same operation twice. Also, the user is allowed to apply the same operation on the local database twice. In the last case, the app will just ignore the operation if it already exists. But the thing is, it will not return any errors (but will print a warning in logs).

It sounds absurd, but if you look at bugs above... I decided to allow it because who knows what will happen in the future. I believe this will help catch future bugs.